[Diffusion][Disagg] Register LTX-2 dual-VAE DAG e2e on 2-GPU CI
Use a colocated 2-GPU layout so the LTX dual-VAE pytest fits the existing
multimodal 2-gpu suite, and move LTX topology stage-coverage checks into unit
tests.
